First two shell scripts
run.sh checks if a given webpage is capable of establishing a connection or not. 

test.sh checks if any change has been detected in a webpage since starting the script.

Both scripts work with BASH. And yes, both run correctly by default on Debian based systems. 

The names of these scripts could've been better. But eh, that's alright